pushed change

+++ flake.nix
@@ -22,6 +22,7 @@
           installPhase = ''
             mkdir -p $out
             cp -r ./* $out/
+            find $out -type f -print0 | sort -z | xargs -0 sha256sum | sha256sum > $out/config.hash
           '';
         };
       in
@@ -38,9 +39,17 @@
           ];
           shellHook = ''
             mkdir -p ~/.config/nvim
-            rsync -av --delete ${nvim-config}/ ~/.config/nvim/
-            echo "Neovim configuration has been updated in ~/.config/nvim"
-            echo "you no need update"
+            if [ -f ~/.config/nvim/config.hash ]; then
+              if diff -q ${nvim-config}/config.hash ~/.config/nvim/config.hash >/dev/null; then
+                echo "Neovim configuration is up to date"
+              else
+                rsync -av --delete ${nvim-config}/ ~/.config/nvim/
+                echo "Neovim configuration has been updated in ~/.config/nvim"
+              fi
+            else
+              rsync -av --delete ${nvim-config}/ ~/.config/nvim/
+              echo "Neovim configuration has been initialized in ~/.config/nvim"
+            fi
           '';
         };
 
@@ -49,8 +58,17 @@
           paths = [
             (pkgs.writeShellScriptBin "setup-nvim" ''
               mkdir -p ~/.config/nvim
-              rsync -av --delete ${nvim-config}/ ~/.config/nvim/
-              echo "Neovim configuration has been updated in ~/.config/nvim"
+              if [ -f ~/.config/nvim/config.hash ]; then
+                if diff -q ${nvim-config}/config.hash ~/.config/nvim/config.hash >/dev/null; then
+                  echo "Neovim configuration is up to date"
+                else
+                  rsync -av --delete ${nvim-config}/ ~/.config/nvim/
+                  echo "Neovim configuration has been updated in ~/.config/nvim"
+                fi
+              else
+                rsync -av --delete ${nvim-config}/ ~/.config/nvim/
+                echo "Neovim configuration has been initialized in ~/.config/nvim"
+              fi
             '')
           ];
         };
